Fixed Dark mod in Header and Default Video layout
In my Updated Code I fixed the darken mode Nav bar effects and provide the perfect visual appealing experience in both light and in dark mode. With this i also add a default Random video section where every time a random video is able to play when refresh the page.